+++ Projekt/SensorGloveQt/rotationmatrix.h
@@ -6,7 +6,7 @@ class RotationMatrix: public QMatrix3x3
     RotationMatrix(const float* p_values);
 public:
     RotationMatrix();
-    RotationMatrix(double p_Angle, char p_Axis);
+    RotationMatrix(float p_Angle, char p_Axis);
 
     QVector3D operator *(const QVector3D & p_Vector3D);
 };
